+++ src/main/java/xyz/mlhmz/mcserverinformation/coordinatelog/stores/PlayerCountStore.java
@@ -3,5 +3,5 @@ package xyz.mlhmz.mcserverinformation.coordinatelog.stores;
 import java.util.UUID;
 
 public interface PlayerCountStore {
-    long incrementAndGetCount(UUID player);
+    int incrementAndGetCount(UUID player);
 }
